Adding grpc server in auth
