+++ horusec-auth/config/admin/admin.go
@@ -1,0 +1,39 @@
+package admin
+
+import (
+	"github.com/ZupIT/horusec/development-kit/pkg/databases/relational"
+	"github.com/ZupIT/horusec/development-kit/pkg/databases/relational/repository/account"
+	accountEntities "github.com/ZupIT/horusec/development-kit/pkg/entities/account"
+	"github.com/ZupIT/horusec/development-kit/pkg/utils/logger"
+	"github.com/ZupIT/horusec/horusec-auth/config/app"
+)
+
+func CreateApplicationAdmin(config *app.Config, read relational.InterfaceRead, write relational.InterfaceWrite) {
+	if config.GetEnableApplicationAdmin() {
+		err := account.NewAccountRepository(read, write).Create(getDefaultAccountApplicationAdmin(config).SetAccountData())
+		if err != nil {
+			if err.Error() != "pq: duplicate key value violates unique constraint \"accounts_email_key\"" {
+				logger.LogPanic("Some error occurs when create application admin", err)
+			} else {
+				logger.LogInfo("Application admin already exists")
+			}
+		} else {
+			logger.LogInfo("Application admin created with success")
+		}
+	}
+}
+
+func getDefaultAccountApplicationAdmin(config *app.Config) *accountEntities.Account {
+	entity, err := config.GetApplicationAdminData()
+	if err != nil {
+		logger.LogPanic("Some error occurs when parse Application Admin Data to Account", err)
+	}
+	pass := entity.Password
+	return &accountEntities.Account{
+		Email:              entity.Email,
+		Password:           pass,
+		Username:           entity.Username,
+		IsConfirmed:        true,
+		IsApplicationAdmin: true,
+	}
+}
